brain/audio/requirements.txt
paho-mqtt>=1.6.1
pyyaml>=6.0
dbus-python>=1.2.18
